(MSVC 2003 Xbox 1) Make PCE core Xbox 1-compatible

+++ mednafen/hw_sound/pce_psg/pce_psg.cpp
@@ -373,7 +373,7 @@ PCE_PSG::PCE_PSG(Blip_Buffer *bb_l, Blip_Buffer *bb_r, int want_revision)
          double flub = 1;
 
          if(vl)
-          flub /= pow(2, (double)1 / 4 * vl);                  // ~1.5dB reduction per increment of vl 
+          flub /= powf(2, (double)1 / 4 * vl);                  // ~1.5dB reduction per increment of vl 
 
 	 if(vl == 0x1F)
 	  flub = 0;
